[ClusterOpenIDConnectPreset] Prefactor: extract defaulting into dedicated file
