+++ pkg/apis/settings/v1alpha1/defaults.go
@@ -15,22 +15,9 @@
 package v1alpha1
 
 import (
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 )
 
 func addDefaultingFuncs(scheme *runtime.Scheme) error {
 	return RegisterDefaults(scheme)
-}
-
-// SetDefaults_ClusterOpenIDConnectPreset sets default values for ClusterOpenIDConnectPreset objects.
-func SetDefaults_ClusterOpenIDConnectPreset(obj *ClusterOpenIDConnectPreset) {
-	if obj.Spec.ShootSelector == nil {
-		obj.Spec.ShootSelector = &metav1.LabelSelector{}
-	}
-
-	if obj.Spec.ProjectSelector == nil {
-		obj.Spec.ProjectSelector = &metav1.LabelSelector{}
-	}
-	setDefaults_KubeAPIServerOpenIDConnect(&obj.Spec.Server)
 }
